layout/style/test/test_transitions_and_zoom.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=583219
-->
<head>
  <title>Test for Bug 583219</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <style type="text/css">

  #display {
    transition: margin-left 1s linear;
  }

  </style>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=583219">Mozilla Bug 583219</a>
<p id="display"></p>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 583219 **/

var p = document.getElementById("display");
var cs = getComputedStyle(p, "");
cs.marginLeft;

p.addEventListener("transitionend", TransitionEndHandler);
p.style.marginLeft = "100px";
cs.marginLeft;

SpecialPowers.setFullZoom(window, 2.0)

SimpleTest.waitForExplicitFinish();

function TransitionEndHandler(event) {
  ok(true, "transition has completed");
  is(event.propertyName, "margin-left", "event.propertyName");
  is(cs.marginLeft, "100px", "value of margin-left");
  SpecialPowers.setFullZoom(window, 1.0)
  SimpleTest.finish();
}

</script>
</pre>
</body>
</html>
